Add image of Odin enjoying chicken wings and update index and styles
- Added a new image of Odin enjoying chicken wings.
- Updated the index.html file to include the new image.
- Modified styles.css to style the new image appropriately.
- Modified header and gave it a border and new colors.

Since the last commit, I added a new image of Odin enjoying chicken wings to enhance the visual appeal of the website. I updated the index.html file to include this new image, ensuring it is displayed correctly. Additionally, I modified the styles.css file to style the new image appropriately, making sure it fits well with the overall design. I also updated the header by giving it a border and new colors to improve the aesthetics and make it more visually appealing.